requirements.txt
requests
pandas
google-cloud-bigquery
python-dotenv
pyarrow
pandas-gbq
